# openapi/schema.openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "0.0.1",
        "title": "Iglu API",
        "description": "Iglu API Docs for use in your development"
    },
    "servers": [{ "url": "/" }],
...
    "paths": {
...
        "/api/v1/cache/{tenant}/multipart-nar/{uid}": {
            "post": {
                "description": "Retrieve an upload url for a given upload id (uid)",
                "summary": "Retrieve upload url",
                "tags": ["api/v1/cache", "cachix"],
                "parameters": [
                    {
                        "schema": { "type": "string" },
                        "required": true,
                        "name": "tenant",
                        "in": "path"
                    },
                    {
                        "schema": { "type": "string" },
                        "required": true,
                        "name": "uid",
                        "in": "path"
                    },
                    {
                        "schema": { "type": "number" },
                        "required": true,
                        "name": "partNumber",
                        "in": "query"
                    },
                    {
                        "schema": { "type": "string" },
                        "required": true,
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "schema": { "type": "string" },
                        "required": true,
                        "name": "user-agent",
                        "in": "header"
                    }
                ],
                "requestBody": {
                    "description": "The MD5 of the derivation part you are uploading",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": { "contentMD5": { "type": "string" } },
                                "required": ["contentMD5"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An upload URL for a derivation part. May either be S3 or local Iglu upload URL",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "uploadURL": { "type": "string", "format": "uri" }
                                    },
                                    "required": ["uploadURL"]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Returned if your body or request params were malformed or not in the appropriate format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The key you were using is not recognized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - No authorization, either you didn't provide an authorization header or your header is not in the Bearer format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The requested tenant does not exist on this server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "408": {
                        "description": "Returned if the upload ID is no longer valid (timed out)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given upload ID was not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returned if iglu was unable to handle the request correctly",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}